fix: Session 전달 방식 수정
