+++ app/RootAppLayout.tsx
@@ -2,19 +2,15 @@
 
 import ReactQueryWrapper from 'components/reactQueryWrapper';
 import { Layout } from 'components/RootLayout';
-import { Session } from 'next-auth';
 import { SessionProvider } from 'next-auth/react';
 import React, { PropsWithChildren } from 'react';
 import { ReactQueryDevtools } from 'react-query/devtools';
 import { ToastContainer } from 'react-toastify';
 import { RecoilRoot } from 'recoil';
 
-function RootAppLayout({
-  children,
-  session,
-}: PropsWithChildren<{ session: Session | null }>) {
+function RootAppLayout({ children }: PropsWithChildren<unknown>) {
   return (
-    <SessionProvider session={session}>
+    <SessionProvider>
       <RecoilRoot>
         <ReactQueryWrapper>
           <ToastContainer />
