+++ app/posts/page.tsx
@@ -2,8 +2,6 @@ import { fetchPostsFromFirebase } from '@apis/posts';
 import { use } from 'react';
 import Posts from './posts';
 
-export const revalidate = 60 * 60;
-
 function PostsPage() {
   const posts = use(fetchPostsFromFirebase());
