+++ app/users/layout.tsx
@@ -1,7 +1,7 @@
 import React, { PropsWithChildren } from 'react';
 
 function PostsLayout({ children }: PropsWithChildren<unknown>) {
-  return <div>aaa{children}</div>;
+  return <div>{children}</div>;
 }
 
 export default PostsLayout;
